rename: move files in wrong path
