+++ src/langs/i18n.js
